moved phone methods for the client model into model concern
