--- mana-api.sublime-project
@@ -1,8 +1,0 @@
-{
-	"folders":
-	[
-		{
-			"path": "."
-		}
-	]
-}
